Silence GCC warning about the log level not being covered by the enum values: this is a application-side error level, which is expected to be out of the default range.

+++ bibtex.c
@@ -41,7 +41,7 @@ void bibtex_message_handler (const gchar *log_domain G_GNUC_UNUSED,
 	fprintf (stderr, "%s: ", name);
     }
 
-    switch (log_level) {
+    switch ((int)log_level) {
 
     case BIB_LEVEL_ERROR:
 	fprintf (stderr, "error: %s\n", message);
